:truck: Move jsonrpc types and utils to separate files update variable names
