+++ framework/src/controller/jsonrpc/index.ts
@@ -12,95 +12,5 @@
  * Removal or modification of this copyright notice is prohibited.
  */
 
-import { validator, LiskValidationError } from '@liskhq/lisk-validator';
-
-type ID = string | number | null;
-export type Result = string | number | boolean | object;
-export interface SuccessObject {
-	jsonrpc: string;
-	id: ID;
-	result: Result;
-}
-export interface NotificationObject {
-	jsonrpc: string;
-	method: string;
-	result?: Result;
-}
-export interface JsonRpcError {
-	code: number;
-	message: string;
-	data?: Result;
-}
-export interface ErrorObject {
-	jsonrpc: string;
-	id: ID;
-	error: JsonRpcError;
-}
-export interface RequestObject {
-	readonly id: string | number | null;
-	readonly jsonrpc: string;
-	readonly method: string;
-	readonly params?: object;
-}
-
-export const VERSION = '2.0';
-
-const RequestSchema = {
-	id: 'jsonRPCRequestSchema',
-	type: 'object',
-	required: ['jsonrpc', 'method', 'id'],
-	properties: {
-		jsonrpc: {
-			type: 'string',
-		},
-		method: {
-			type: 'string',
-		},
-		id: {
-			type: ['number', 'string', 'null'], // https://www.jsonrpc.org/specification#id1
-		},
-		params: {
-			type: 'object',
-		},
-	},
-};
-
-export const validateJSONRPC = (data: object): void => {
-	const errors = validator.validate(RequestSchema, data);
-	if (errors.length) {
-		throw new LiskValidationError(errors);
-	}
-};
-
-export const successObject = (id: ID, result: Result): SuccessObject => ({
-	jsonrpc: VERSION,
-	id,
-	result,
-});
-
-export const notificationObject = (method: string, result?: Result): NotificationObject => ({
-	jsonrpc: VERSION,
-	method,
-	result,
-});
-
-export const errorObject = (id: ID, error: JsonRpcError): ErrorObject => ({
-	jsonrpc: VERSION,
-	id,
-	error,
-});
-
-export const invalidRequest = (): JsonRpcError => ({ message: 'Invalid request', code: -32600 });
-
-export const methodNotFound = (): JsonRpcError => ({ message: 'Method not found', code: -32601 });
-
-export const invalidParams = (): JsonRpcError => ({ message: 'Invalid params', code: -32602 });
-
-export const internalError = (data?: Result): JsonRpcError => {
-	if (data) {
-		return { message: 'Internal error', code: -32603, data };
-	}
-	return { message: 'Internal error', code: -32603 };
-};
-
-export const parseError = (): JsonRpcError => ({ message: 'Parse error', code: -32700 });
+export * from './types';
+export * from './utils';
