+++ framework/test/integration/controller/ipc_channel_without_bus.spec.ts
@@ -18,7 +18,7 @@ import { resolve as pathResolve } from 'path';
 import { IPCChannel } from '../../../src/controller/channels';
 import { IPCServer } from '../../../src/controller/ipc/ipc_server';
 import { Event } from '../../../src/controller/event';
-import * as JSONRPC from '../../../src/controller/jsonrpc';
+import * as jsonRPC from '../../../src/controller/jsonrpc';
 
 const socketsDir = pathResolve(`${homedir()}/.lisk/functional/ipc_channel_without_bus/sockets`);
 
@@ -107,7 +107,7 @@ describe('IPCChannel', () => {
 					// Act
 					alphaChannel.subscribe(`${beta.moduleAlias}:${eventName}`, data => {
 						// Assert
-						const eventData = (Event.fromJSONRPC(JSONRPC.notificationObject('app:new:block', data))
+						const eventData = (Event.fromJSONRPC(jsonRPC.notificationObject('app:new:block', data))
 							.result as { data: object }).data;
 						expect(eventData).toEqual(betaEventData);
 						resolve();
@@ -127,7 +127,7 @@ describe('IPCChannel', () => {
 					// Act
 					alphaChannel.once(`${beta.moduleAlias}:${eventName}`, data => {
 						// Assert
-						const eventData = (Event.fromJSONRPC(JSONRPC.notificationObject('app:new:block', data))
+						const eventData = (Event.fromJSONRPC(jsonRPC.notificationObject('app:new:block', data))
 							.result as { data: object }).data;
 						expect(eventData).toEqual(betaEventData);
 						resolve();
@@ -150,7 +150,7 @@ describe('IPCChannel', () => {
 					// Act
 					betaChannel.once(`${alpha.moduleAlias}:${eventName}`, data => {
 						// Assert
-						const eventData = (Event.fromJSONRPC(JSONRPC.notificationObject('app:new:block', data))
+						const eventData = (Event.fromJSONRPC(jsonRPC.notificationObject('app:new:block', data))
 							.result as { data: object }).data;
 						expect(eventData).toEqual(alphaEventData);
 						done();
